DSL: Fix `RouterSpec` for several subflows
The root if issue is up to the order of bean registration in the beanFactory.
In this case the first subflows is registered correctly, but the next on has been registered after `RouterSubFlowMappingProvider`.
In this case the input `channel` of the next subflow hasn't been registered yet, hence `NPE` in the `this.router.setChannelMapping`
from `@PostConstruct` of `RouterSubFlowMappingProvider`.

Move the `RouterSubFlowMappingProvider` to end of `commentsToRegister` collection to give a chance to register all subflows before the `RouterSubFlowMappingProvider`